Refactor first, then the rest.
