+++ src/main/kotlin/dev/fidil/fettle/FettleApplication.kt
@@ -2,8 +2,11 @@
 
 package dev.fidil.fettle
 
+import dev.fidil.fettle.command.BranchProtectionCommand
+import dev.fidil.fettle.command.DependabotCommand
 import kotlinx.cli.ArgParser
 import kotlinx.cli.ExperimentalCli
+import java.util.*
 import kotlin.system.exitProcess
 
 
@@ -22,5 +25,3 @@ fun main(args: Array<String>) {
     println("╭ᥥ╮(´• ᴗ •`˵)╭ᥥ╮")
 
 }
-
-
